requirements.txt
asgiref==3.4.1
Django==3.2.6
djangorestframework==3.12.4
Faker==8.12.0
mysqlclient==2.0.1
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
typing-extensions==3.10.0.0
